# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: Aqar Data Studio — Professional Real Estate Data Management Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /regions:
    get:
      operationId: listRegions
      tags: [regions]
      summary: List all regions
      responses:
        "200":
          description: Region list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Region"

    post:
      operationId: createRegion
      tags: [regions]
      summary: Create a new region
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegionInput"
      responses:
        "201":
          description: Created region
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Region"
  /regions/{id}:
    patch:
      operationId: updateRegion
      tags: [regions]
      summary: Update a region
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegionUpdate"
      responses:
        "200":
          description: Updated region
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Region"

    delete:
      operationId: deleteRegion
      tags: [regions]
      summary: Delete a region
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"

  # ─── PROPERTY TYPES ─────────────────────────────────────────────────────────
  /property-types:
    get:
      operationId: listPropertyTypes
      tags: [property-types]
      summary: List all property types
      responses:
        "200":
          description: Property type list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PropertyType"

    post:
      operationId: createPropertyType
      tags: [property-types]
      summary: Create a new property type
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PropertyTypeInput"
      responses:
        "201":
          description: Created property type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PropertyType"
  /property-types/{id}:
    patch:
      operationId: updatePropertyType
      tags: [property-types]
      summary: Update a property type
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PropertyTypeUpdate"
      responses:
        "200":
          description: Updated property type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PropertyType"

    delete:
      operationId: deletePropertyType
      tags: [property-types]
      summary: Delete a property type
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"

  # ─── LOOKUP OPTIONS (Dynamic) ─────────────────────────────────────────────
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"

    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"

  # ─── AUDIT LOGS ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    DeleteResult:
      type: object
      required: [success]
      properties:
        success: { type: boolean }
        id: { type: string }

    # ── Property ────────────────────────────────────────────────────────────
...
    Region:
      type: object
      required: [id, name, active]
      properties:
        id: { type: string }
        name: { type: string }
        active: { type: boolean }
        propertyCount: { type: number }
    RegionInput:
      type: object
      required: [id, name]
      properties:
        id: { type: string }
        name: { type: string }
        active: { type: boolean }
    RegionUpdate:
      type: object
      properties:
        name: { type: string }
        active: { type: boolean }

    # ── PropertyType ──────────────────────────────────────────────────────
    PropertyType:
      type: object
      required: [id, name, active]
      properties:
        id: { type: string }
        name: { type: string }
        active: { type: boolean }
        propertyCount: { type: number }
    PropertyTypeInput:
      type: object
      required: [id, name]
      properties:
        id: { type: string }
        name: { type: string }
        active: { type: boolean }
    PropertyTypeUpdate:
      type: object
      properties:
        name: { type: string }
        active: { type: boolean }

    # ── LookupOption (dynamic values) ─────────────────────────────────────
...
    User:
      type: object
      required: [id, email, role, active, createdAt]
      properties:
        id: { type: string }
        email: { type: string }
        name: { type: ["string", "null"] }
        role: { type: string, enum: [admin, agent, viewer] }
        active: { type: boolean }
        createdAt: { type: string }
        lastSignIn: { type: ["string", "null"] }
    UserInput:
      type: object
      required: [email, role]
      properties:
        email: { type: string, format: string }
        name: { type: string }
        role: { type: string, enum: [admin, agent, viewer] }
        active: { type: boolean }
    UserUpdate:
      type: object
      properties:
        name: { type: string }
        role: { type: string, enum: [admin, agent, viewer] }
        active: { type: boolean }

    # ── Audit Log ─────────────────────────────────────────────────────────
...