# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Aqar Data Studio — Professional Real Estate Data Management Platform
 * OpenAPI spec version: 1.0.0
 */
import type { UserRole } from './userRole';

export interface User {
  id: string;
  email: string;
  /** @nullable */
  name?: string | null;
  role: UserRole;
  active: boolean;
  createdAt: string;
  /** @nullable */
  lastSignIn?: string | null;
}